+++ CHANGELOG.md
@@ -14,6 +14,7 @@ backwards compatibility.
 * Make it possible to offset scale steps by up to 4 quarter notes in either
   direction.
 * Allow using a custom "noise" burst source.
+* Make it possible to control the number of strummed strings.
 
 ## 0.1.1